Reorganize deployer sort tests
